Moved healthcheck to utils package
